SBERDOMA-1005 Move file to containers && PoC
